requirements.txt
numpy
opencv-python
matplotlib
scikit-learn
tensorflow
flask
mysql-connector-python
